+++ src/App.js
@@ -3,6 +3,7 @@ import Root from "./Root.js";
 import Company from "./components/Main/Company/Company.jsx";
 import Metrics, { thresholdsLoader } from "./components/Main/Metrics/Metrics.jsx";
 import Selector, { filtersLoader } from "./components/Main/Selector/Selector.jsx";
+import SelectorContextProvider from "./store/selector-context.jsx";
 import User from "./components/Main/User/User.jsx";
 
 import "./App.css";
@@ -13,7 +14,7 @@ const router = createBrowserRouter([
     element: <Root />,
     children: [
       { index: true, element: <Metrics />, loader: thresholdsLoader },
-      { path: '/selector', element: <Selector />, loader: filtersLoader},
+      { path: '/selector', element: <SelectorContextProvider><Selector /></SelectorContextProvider>, loader: filtersLoader},
       { path: '/company', element: <Company />},
       { path: '/user', element: <User />},
     ],