+++ src/components/Main/Selector/Selector.jsx
@@ -1,29 +1,30 @@
+import { useContext, useEffect } from "react";
 import { useLoaderData } from "react-router-dom";
 import SelectorBuilder from "./SelectorBuilder/SelectorBuilder.jsx";
 import { SelectorContext } from "../../../store/selector-context.jsx";
 import $Selector from "./Selector.styles.jsx";
 
 export default function Selector() {
-  const selectorFilters = useLoaderData();
+  const selectorFiltersData = useLoaderData();
 
-  const ctxSelector = {
-    currentSelection: [{ command: "show", filter: "area" }, { command: "by", filter: "area"}],
-    selectionHandler: () => {},
-    selectorFilters: selectorFilters
-  };
+  const { setSelectorFilters } = useContext(SelectorContext);
+
+  useEffect(() => {
+    setSelectorFilters(selectorFiltersData);
+  }, []);
 
   return (
-    <SelectorContext.Provider value={ctxSelector}>
+    <>
       <$Selector>
         <SelectorBuilder />
       </$Selector>
       charts elements
-    </SelectorContext.Provider>
+    </>
   );
 }
 
 export async function filtersLoader() {
-  const response = await fetch('http://localhost:8080/filter/filters');
+  const response = await fetch("http://localhost:8080/filter/filters");
   const resData = await response.json();
 
   if (!response.ok) {