+++ src/components/Main/Selector/SelectorBuilder/SelectorBuilder.jsx
@@ -1,12 +1,15 @@
 import $SelectorBuilder from "./SelectorBuilder.styles.jsx";
 import BuilderButton from "./BuilderButton/BuilderButton.jsx";
 import BuilderFilter from "./BuilderFilter/BuilderFilter.jsx";
+import BuilderLabel from "./BuilderLabel/BuilderLabel.jsx";
 
 export default function SelectorBuilder() {
   return (
     <$SelectorBuilder>
       <BuilderButton />
-      <BuilderFilter />
+      <BuilderFilter criteria='show'/>
+      <BuilderLabel />
+      <BuilderFilter criteria='where'/>
     </$SelectorBuilder>
   );
 }